pyproject.toml
[project]
name = "crossfit-timetable"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "lxml>=6.0.2",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.0.0",
    "python-dateutil>=2.9.0.post0",
    "fastapi[standard]>=0.115.0",
    "icalendar>=5.0.0",
    "aiohttp>=3.10.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
package = true
